refactor: Header contents 분리
