+++ pages/_app.tsx
@@ -18,7 +18,7 @@ export default function CustomApp({ Component, pageProps }: AppProps) {
         },
       })
   )
-  const isToken = useLogin()
+  // const isToken = useLogin()
 
   return (
     <>
